refactor(mfa): merge MFA page into Password & Auth, split out TOTP module
Move two-factor auth onto the password page, which becomes 'Password &
Auth': TOTP lives in its own totp.tsx module, the smaller read-only OAuth
card lives inline in password.tsx, both gated to the server build. The
standalone MFA page is marked for deletion (enforceDeleted) and removed
from the content widget registry. Drop the page-level status badge and
the MFA intro blurb, and prune the translation keys left unused.
