+++ apps/client/src/widgets/type_widgets/options/password.css
@@ -1,0 +1,17 @@
+/* OAuth section heading: the title text and its configured/not-configured status badge on one line,
+   the badge sitting just after the title. */
+.oauth-status-title {
+    display: inline-flex;
+    align-items: center;
+    gap: 10px;
+}
+
+/* Status badge mirrors the page-header MFA badge: a colored outline whose hue is driven by --color
+   (green when configured, muted otherwise). */
+.oauth-status-badge.configured {
+    --color: var(--admonition-tip-accent-color, #40c025);
+}
+
+.oauth-status-badge.not-configured {
+    --color: var(--muted-text-color, #888);
+}
