+++ apps/client/src/widgets/type_widgets/options/totp.css
@@ -1,43 +1,3 @@
-/* Flex row holding the status badge, so the badge sizes to its content instead of stretching to
-   fill the full header width (the shared actions slot isn't itself a flex container). */
-.mfa-header-actions {
-    display: flex;
-    align-items: center;
-}
-
-/* Status badge on the MFA page header, mirroring the shortcuts page's outline badge: a colored
-   outline whose hue is driven by the --color variable (green when active, muted when inactive). */
-.mfa-status-badge {
-    flex: 0 0 auto;
-    margin-inline-end: auto;
-}
-
-.mfa-status-badge.active {
-    --color: var(--admonition-tip-accent-color, #40c025);
-}
-
-.mfa-status-badge.inactive {
-    --color: var(--muted-text-color, #888);
-}
-
-/* OAuth section heading: the title text and its configured/not-configured status badge on one line,
-   the badge sitting just after the title. */
-.oauth-status-title {
-    display: inline-flex;
-    align-items: center;
-    gap: 10px;
-}
-
-/* Status badge mirrors the page-header MFA badge: a colored outline whose hue is driven by --color
-   (green when configured, muted otherwise). */
-.oauth-status-badge.configured {
-    --color: var(--admonition-tip-accent-color, #40c025);
-}
-
-.oauth-status-badge.not-configured {
-    --color: var(--muted-text-color, #888);
-}
-
 /* Keeps the "Recovery codes" label text and its dots on one line, the dots sitting just after the
    title. Lives inside the row's <label>, so text and dots share one box and stay aligned. */
 .recovery-codes-title {
