+++ apps/server/src/assets/translations/en/server.json
@@ -287,7 +287,7 @@
     "spellcheck-title": "Spellcheck",
     "desktop-title": "Desktop",
     "security-title": "Security",
-    "password-title": "Password",
+    "password-title": "Password & Auth",
     "multi-factor-authentication-title": "MFA",
     "etapi-title": "ETAPI",
     "backup-title": "Backup",
